--- src/main/java/baseball/Computer.java
@@ -1,22 +1,0 @@
-package baseball;
-
-import java.util.List;
-
-public class Computer {
-    private RandomNumbers randomNumbers;
-
-    public Computer() {
-    }
-
-    public void init() {
-        randomNumbers = new RandomNumbers();
-    }
-
-    public int checkBallCount(List<Integer> playerNumbers) {
-        return randomNumbers.checkBallCount(playerNumbers);
-    }
-
-    public int checkStrikeCount(List<Integer> playerNumbers) {
-        return randomNumbers.checkStrikeCount(playerNumbers);
-    }
-}
