+++ src/main/java/baseball/domain/Player.java
@@ -1,14 +1,16 @@
-package baseball;
+package baseball.domain;
 
 import static baseball.utils.Constants.TOTAL_BALL_COUNTS;
 import static baseball.utils.GameMessage.GAME_SUCCESS_MESSAGE;
 import static baseball.utils.GameMessage.INPUT_REQUEST_MESSAGE;
 
+import baseball.domain.wrapped.Ball;
+import baseball.domain.wrapped.Strike;
 import baseball.view.InputView;
 import baseball.view.OutputView;
-import camp.nextstep.edu.missionutils.Console;
 
 public class Player {
+    private RandomNumbers randomNumbers = new RandomNumbers();
     private Computer computer = new Computer();
     private Strike strike;
     private Ball ball;
@@ -25,19 +27,15 @@ public class Player {
     public void play() {
         PlayerNumbers playerNumbers;
         while (strike.isLessThan(TOTAL_BALL_COUNTS)) {
-            OutputView.printMessage(INPUT_REQUEST_MESSAGE);
+            OutputView.printMessageWithoutNewLine(INPUT_REQUEST_MESSAGE);
             String playerInput = InputView.getPlayerInput();
             playerNumbers = new PlayerNumbers(playerInput);
 
-            strike = new Strike(playerNumbers.checkStrikeCount(computer));
-            ball = new Ball(playerNumbers.checkBallCount(computer));
+            strike = new Strike(playerNumbers.checkStrikeCount(randomNumbers));
+            ball = new Ball(playerNumbers.checkBallCount(randomNumbers));
             printBallCount();
         }
-        printGameSuccessMessage();
-    }
-
-    private void printGameSuccessMessage() {
-        System.out.println(GAME_SUCCESS_MESSAGE);
+        OutputView.printMessage(GAME_SUCCESS_MESSAGE);
     }
 
     public void printBallCount() {
@@ -54,6 +52,6 @@ public class Player {
         if (!ball.getResultMessage().equals("0볼") && !strike.getResultMessage().equals("0스트라이크")) {
             result = ball.getResultMessage() + " " + strike.getResultMessage();
         }
-        System.out.println(result);
+        OutputView.printMessage(result);
     }
 }
