+++ src/main/java/baseball/domain/wrapped/Ball.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain.wrapped;
 
 public class Ball {
     private int ball;
