+++ src/main/java/baseball/domain/wrapped/Strike.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain.wrapped;
 
 public class Strike {
     private int strike;
